+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2023-08-04
+
+- [#36](https://github.com/itk-dev/devops_itksites/pull/36)
+  Implemented OIDC code flow and handled target path after login.
+
 ## [1.4.0] - 2023-08-01
 
 - [#34](https://github.com/itk-dev/devops_itksites/pull/34)
@@ -38,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-09-15
 
-[Unreleased]: https://github.com/itk-dev/devops_itksites/compare/1.4.0...HEAD
+[Unreleased]: https://github.com/itk-dev/devops_itksites/compare/1.4.1...HEAD
+[1.4.1]: https://github.com/itk-dev/devops_itksites/compare/1.4.0...1.4.1
 [1.4.0]: https://github.com/itk-dev/devops_itksites/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/itk-dev/devops_itksites/compare/1.2.2...1.3.0
 [1.2.2]: https://github.com/itk-dev/devops_itksites/compare/1.2.1...1.2.2